Load SAMAudio in float16 to halve main model memory (~6GB -> ~3GB)
Total sub-component footprint (ImageBind 4.5GB + T5 0.9GB + RoBERTa 0.5GB
+ ranker 1.9GB) leaves little headroom in a 15GB container; casting the
main model to fp16 before moving to device keeps total RAM under the limit.

+++ scripts/run_text_prompt.py
@@ -37,7 +37,7 @@ def main() -> None:
     print(f"Loading model from {model_dir} ...")
 
     from sam_audio import SAMAudio, SAMAudioProcessor
-    model = SAMAudio.from_pretrained(model_dir, local_files_only=True).to(device).eval()
+    model = SAMAudio.from_pretrained(model_dir, local_files_only=True).half().to(device).eval()
     processor = SAMAudioProcessor.from_pretrained(model_dir, local_files_only=True)
 
     print(f"Audio:  {args.audio}")
